修复录制视频后视频帧播放过快的问题

+++ libmedia/src/main/cpp/component/encoder/FFBaseEncoder.h
@@ -117,7 +117,7 @@ public:
 
 
     double GetCurrentTimestamp() {
-        return m_NextPts * av_q2d(m_CodecCtx->time_base);
+        return (double) m_NextPts * av_q2d(m_CodecCtx->time_base);
     }
 
     ~FFBaseEncoder() {
